requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
authlib==1.3.0
httpx==0.24.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
pydantic==1.10.10
python-jose==3.3.0
psycopg2-binary==2.9.7
jinja2==3.1.2
python-multipart==0.0.6
itsdangerous
bcrypt>=4.0.0
passlib[bcrypt]>=1.7.4


